backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
open3d>=0.18.0
pyvista>=0.43.0
cadquery>=2.4.0
anthropic>=0.25.0
pydantic>=2.6.0
python-multipart>=0.0.9
pillow>=10.0.0
numpy>=1.26.0
